core: allow data sorting for unordered streams
A production ULog in Aviant failed comparison with the original ULog
after write_ulog, which turned out to be because of messages arriving
at different times and thus the timestamps weren't completely in order.
This isn't really a problem, so we want the tests to pass, but in order
to do that we need the option of sorting the datasets.